server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.18
alpha-vantage==2.3.1
python-binance==1.0.19
ta==0.10.2
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.25.2
websockets==12.0

# Database drivers
motor==3.1.1  # MongoDB async driver (compatible version)
pymongo==4.3.3  # MongoDB driver (compatible with motor 3.1.1)
psycopg2-binary==2.9.7  # PostgreSQL
asyncpg==0.29.0  # PostgreSQL async driver
pymysql==1.1.0  # MySQL
aiomysql==0.2.0  # MySQL async driver
sqlalchemy==2.0.23  # SQL ORM
alembic==1.12.1  # Database migrations 